Move YearlyMetricsHelper impl to a separate file
